kern - Properly return the number of bytes in hw.physmem sysctl.
- Changed physmem from int to long, now it can hold a
  really high number of pages.
- Return an unsigned long value with appropiate formatting
  for hw.physmem.
- Add description for it.

Based-upon: FreeBSD

+++ sys/platform/vkernel64/platform/init.c
@@ -83,7 +83,7 @@
 vm_paddr_t phys_avail[16];
 vm_paddr_t Maxmem;
 vm_paddr_t Maxmem_bytes;
-int physmem;
+long physmem;
 int MemImageFd = -1;
 struct vkdisk_info DiskInfo[VKDISK_MAX];
 int DiskNum;
